QNX: Fixed running the application if first attempt failed
When the first attempt to run an application on the BlackBerry device failed,
e.g. due to not being able to connect to the device, the Play button remained
disabled, so any subsequent attempts to run the application was not possible.

+++ src/plugins/qnx/blackberryapplicationrunner.cpp
@@ -155,6 +155,7 @@ void BlackBerryApplicationRunner::startFinished(int exitCode, QProcess::ExitStat
         const QString errorString = (m_launchProcess->error() != QProcess::UnknownError)
                 ? m_launchProcess->errorString() : tr("Launching application failed");
         emit startFailed(errorString);
+        reset();
     }
 }
 
@@ -344,9 +345,11 @@ void BlackBerryApplicationRunner::reset()
     m_stopping = false;
 
     m_runningStateTimer->stop();
-    m_runningStateProcess->terminate();
-    if (!m_runningStateProcess->waitForFinished(1000))
-        m_runningStateProcess->kill();
+    if (m_runningStateProcess) {
+        m_runningStateProcess->terminate();
+        if (!m_runningStateProcess->waitForFinished(1000))
+            m_runningStateProcess->kill();
+    }
 
     if (m_tailProcess && m_tailProcess->isProcessRunning())
         killTailProcess();
